Update project names

+++ MedicalDocumentationManager.Domain/Implementation/FileLoggerFactory.cs
@@ -1,4 +1,5 @@
 ﻿using MedicalDocumentationManager.Domain.Abstraction;
+using MedicalDocumentationManager.Domain.Abstraction.Contracts;
 
 namespace MedicalDocumentationManager.Domain.Implementation;
 
@@ -8,15 +9,9 @@ public class FileLoggerFactory : ILoggerFactory
 
     public FileLoggerFactory(string? filePath)
     {
-        if (string.IsNullOrEmpty(filePath))
-        {
-            filePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "log.txt");
-        }
-        
-        if (!File.Exists(filePath))
-        {
-            File.Create(filePath).Dispose();
-        }
+        if (string.IsNullOrEmpty(filePath)) filePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "log.txt");
+
+        if (!File.Exists(filePath)) File.Create(filePath).Dispose();
 
         _filePath = filePath;
     }
